% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-missing-datetimes.R
\name{ps_add_missing_datetimes}
\alias{ps_add_missing_datetimes}
\title{Add Missing SequenceDate Times.}
\usage{
ps_add_missing_datetimes(x, datetime = "DateTime", by = character(0))
}
\arguments{
\item{x}{The data.frame.}

\item{datetime}{A string naming the POSIXt column.}

\item{by}{A character vector of columns to add the missing datetimes by.}
}
\value{
A tbl data frame sorted by DateTime.
}
\description{
Adds missing datetimes to a data frame.
All missing datetimes between the first and last datetime
with the smallest interval
in the data are added.
}
