% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-datas.R
\name{ps_bind_datas}
\alias{ps_bind_datas}
\title{Bind Data}
\usage{
ps_bind_datas(envir = parent.frame(), rm = TRUE)
}
\arguments{
\item{envir}{The environment containing the data frames.}

\item{rm}{A flag indicating wether to remove the data frames.}
}
\value{
The bound data frames as tbl data frame or an empty data frame.
}
\description{
Binds all the data frames in the environment by rows (and removes them)
}
