% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-files.R
\name{ps_bind_files}
\alias{ps_bind_files}
\title{Bind Data}
\usage{
ps_bind_files(dir = ".", pattern = "[.]csv$", recursive = FALSE,
  read = readr::read_csv, ...)
}
\arguments{
\item{dir}{A string of the directory.}

\item{pattern}{A string of the regular expression to use to identify files.}

\item{recursive}{A flag indicating whether to include files in subdirectories.}

\item{read}{A function to read the files.}

\item{...}{Additional arguments passed to \code{read}.}
}
\value{
The bound data frames as tbl data frame.
}
\description{
Binds all the data frames in files by rows (and removes them)
}
