% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative-sequence.R
\name{ps_cumulative_sequence}
\alias{ps_cumulative_sequence}
\title{Cumulative Sequence}
\usage{
ps_cumulative_sequence(x, cumsum, data, sequence = "DateTime",
  value = "Value")
}
\arguments{
\item{x}{A scalar specifying the start.}

\item{cumsum}{A scalar specifying the cumulative sum.}

\item{data}{The data.frame.}

\item{sequence}{A string naming the sequence column.}

\item{value}{A string of the value column.}
}
\value{
A scalar of the end.
}
\description{
Determines when in sequence a cumulative sum is reached.
Useful for ATUs.
}
