% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-to-numeric.R
\name{ps_cut_to_numeric}
\alias{ps_cut_to_numeric}
\title{Cut to Numeric}
\usage{
ps_cut_to_numeric(x)
}
\arguments{
\item{x}{A character or factor vector of cuts}
}
\value{
An numeric vector of the cuts.
}
\description{
Cut to Numeric
}
\examples{
ps_cut_to_numeric(c(NA, "(0.25,0.75]", "[-2,-1)"))
}
