% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-map.R
\name{ps_env_map_df}
\alias{ps_env_map_df}
\title{Environment map}
\usage{
ps_env_map_df(fun, envir = parent.frame())
}
\arguments{
\item{fun}{A function.}

\item{envir}{The environment.}
}
\value{
An invisible vector of the modified data frames.
}
\description{
Map a function to all data.frames in envir.
}
