% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{ps_get_group}
\alias{ps_get_group}
\title{Get Group}
\usage{
ps_get_group(
  x,
  breaks = c(Recruit = 0, Juvenile = 20, Subadult = 50, Adult = 70)
)
}
\arguments{
\item{x}{A numeric or integer vector}

\item{breaks}{A uniquely named numeric vector of the first value of each group.}
}
\value{
An ordered factor of the group.
}
\description{
Get Group
}
\examples{
ps_get_group(c(-Inf,-1,0,19,20,69,70,Inf))
}
