% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate-sequence.R
\name{ps_interpolate_sequence}
\alias{ps_interpolate_sequence}
\title{Interpolate Sequence}
\usage{
ps_interpolate_sequence(
  x,
  sequence = "DateTime",
  value = "Value",
  by = character(0),
  max_gap = 10L,
  method = "linear",
  step = 0.5
)
}
\arguments{
\item{x}{The data.frame.}

\item{sequence}{A string naming the sequence column.}

\item{value}{A character vector of the value column(s).}

\item{by}{A character vector of columns to interpolate by.}

\item{max_gap}{A count of the maximum gap to interpolate within.}

\item{method}{A string specifying the method (linear or constant).}

\item{step}{An probability (number between 0 and 1 inclusive) specifing
the compromise between left- and right- continuous step function.}
}
\value{
A tbl data frame sorted by sequence and by.
}
\description{
Interpolates missing values in sequence using \code{\link[stats:approx]{stats::approx()}}
}
