% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-excel.R
\name{ps_load_excel}
\alias{ps_load_excel}
\title{Load Excel sheets}
\usage{
ps_load_excel(file = ".", rename = identity, envir = parent.frame())
}
\arguments{
\item{file}{A string of the path to the file.}

\item{rename}{A function that is used to rename files (after removing extension .csv) before they are passed to make.names.}

\item{envir}{The environment to assign the data frames.}
}
\value{
An invisible character vector of the sheet names.
}
\description{
Loads all excel sheets in a workbook as data.frames into specified environment.
}
