% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-files.R
\name{ps_load_files}
\alias{ps_load_files}
\title{Load Files}
\usage{
ps_load_files(dir = ".", pattern = "[.]csv$", recursive = FALSE,
  read = readr::read_csv, add_name = NULL, rename = identity,
  envir = parent.frame(), ...)
}
\arguments{
\item{dir}{A string of the directory.}

\item{pattern}{A string of the regular expression to use to identify files.}

\item{recursive}{A flag indicating whether to include files in subdirectories.}

\item{read}{A function to read the files.}

\item{add_name}{A string specifying the name of the column to add to each data frame with the file names (including subdirectories).}

\item{rename}{A function that is used to rename files (after removing the extension) before they are passed to \code{make.names}.}

\item{envir}{The environment to assign the data frames.}

\item{...}{Additional arguments passed to \code{read}.}
}
\value{
An invisible character vector of the file names.
}
\description{
By default reads all files in the directory dir that match the pattern
using \code{read_csv} and loads them into the
calling environment after renaming using \code{make.names(unique = TRUE)}.
It is useful for quickly loading the data in a set of files into the workspace
(just call the function at the top level of a script).
}
