% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names-datas.R
\name{ps_names_datas}
\alias{ps_names_datas}
\title{Names Data Frames}
\usage{
ps_names_datas(envir = parent.frame())
}
\arguments{
\item{envir}{The environment to get the names of the data frames for.}
}
\value{
A character vector of the names of the data frames.
}
\description{
Gets names of the data frames in by default the calling environment.
}
\examples{
ps_names_datas()
}
