% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-variable.R
\name{ps_select_variable}
\alias{ps_select_variable}
\title{Select Variable Columns}
\usage{
ps_select_variable(x, na_ignore = FALSE)
}
\arguments{
\item{x}{A data frame.}

\item{na_ignore}{A flag specifying whether to ignore missing values when
determining if variable.}
}
\value{
A data frame with constant columns removed.
}
\description{
Select Variable Columns
}
\examples{
ps_select_variable(data.frame(x = c(1, 1), y = c(2, 1)))
}
