% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{ps_write_data_csvs}
\alias{ps_write_data_csvs}
\title{Write Data}
\usage{
ps_write_data_csvs(dir = ".", envir = parent.frame())
}
\arguments{
\item{dir}{A string of the directory.}

\item{envir}{The environment to find the data frames.}
}
\value{
An invisible character vector of the object names.
}
\description{
Writes all the data frames as files in the directory dir
using \code{write_csv}.
}
