% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{size_gaps}
\alias{size_gaps}
\title{Size Gaps}
\usage{
size_gaps(x)
}
\arguments{
\item{x}{A logical vector where TRUE indicates a gap.}
}
\value{
An integer vector of the size of gap each element occurs in.
}
\description{
Size Gaps
}
\keyword{internal}
