% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{poisdata}
\alias{poisdata}
\alias{package-poisdata}
\alias{poisdata-package}
\title{poisdata}
\description{
Read and write data for Poisson Consulting scripts and packages.
}
\seealso{
\url{https://github.com/poissonconsulting/poisdata}
}
