% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-csvs.R
\name{ps_load_csvs}
\alias{ps_load_csvs}
\title{Load CSVs}
\usage{
ps_load_csvs(dir = ".", pattern = "[.]csv$", recursive = FALSE,
  rename = identity, envir = parent.frame(), ...)
}
\arguments{
\item{dir}{A string of the directory.}

\item{pattern}{A string of the regular expression to use to identify files.}

\item{recursive}{A flag indicating whether to include files in subdirectories.}

\item{rename}{A function that is used to rename files (after removing extension .csv) before they are passed to make.names.}

\item{envir}{The environment to assign the data frames.}

\item{...}{Additional arguments passed to \code{read_csv}.}
}
\value{
An invisible character vector of the file names.
}
\description{
Load CSVs
}
