% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-csvs.R
\name{ps_load_csvs}
\alias{ps_load_csvs}
\title{Load CSVs}
\usage{
ps_load_csvs(
  dir = ".",
  recursive = FALSE,
  add_name = NULL,
  rename = identity,
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{dir}{A string of the directory.}

\item{recursive}{A flag indicating whether to include files in subdirectories.}

\item{add_name}{A string specifying the name of the column to add to each data frame with the file names (including subdirectories).}

\item{rename}{A function that is used to rename files (after removing the extension) before they are passed to \code{make.names}.}

\item{envir}{The environment to assign the data frames.}

\item{...}{Additional arguments passed to \code{read}.}
}
\value{
An invisible character vector of the file names.
}
\description{
Load CSVs
}
